Set unix_closedir_job C binding in its own C file
