+++ src/unix/jbuild
@@ -185,6 +185,7 @@
     unix_utimes_job
     unix_isatty_job
     unix_opendir_job
+    unix_closedir_job
     not_available
   ))
   (install_c_headers (
