+++ src/unix/windows_c/not_available.c
@@ -68,4 +68,5 @@ LWT_NOT_AVAILABLE1(unix_fstat_64_job)
 LWT_NOT_AVAILABLE3(unix_utimes_job)
 LWT_NOT_AVAILABLE1(unix_isatty_job)
 LWT_NOT_AVAILABLE1(unix_opendir_job)
+LWT_NOT_AVAILABLE1(unix_closedir_job)
 #endif
